requirements.txt
streamlit
pyspark
scikit-learn
joblib
dbt-core
apache-airflow
pandas
delta-spark
